WIP after coding session:
 * Removed hardcoded "email.h" from dispatcher
 * Moved types and includes for crt.c to crt.h
 * Removed header file generation from C backend
 * Removed trailing whitespace
 * Moved main from crt to seperate wrapper file (not working yet)
