+++ crt/crt.h
@@ -1,0 +1,79 @@
+#include <stdlib.h>
+#include <stdint.h>
+#include <stdbool.h>
+#include <string.h>
+#include <inttypes.h>
+#include <getopt.h>
+#include <unistd.h>
+#include <sys/time.h>
+#include <stdio.h>
+#ifdef __APPLE__
+#   include <mach-o/dyld.h>
+#endif
+
+#define RETC_PRINT_USAGE     1
+#define RETC_PRINT_INFO      2
+#define RETC_PRINT_TABLE     3
+
+#define OUTBUFFER_SIZE       (16*1024)
+#define INBUFFER_SIZE        (16*1024)
+#define INITIAL_BUFFER_SIZE  (4096*8)
+#define OUTBUFFER_STACK_SIZE (1024)
+
+#ifdef FLAG_NOINLINE
+#define INLINE static
+#endif
+#ifndef INLINE
+#define INLINE static inline
+#endif
+
+#ifndef BUFFER_UNIT_T
+#warning "BUFFER_UNIT_T not defined. Falling back to default 'uint8_t'"
+#define BUFFER_UNIT_T uint8_t
+#endif
+
+#ifndef NUM_PHASES
+#warning "NUM_PHASES not defined."
+#endif
+
+#ifndef OUTSTREAM
+#define OUTSTREAM stdout
+#endif
+
+// Order of descriptors provided by pipe()
+#define  READ_FD 0
+#define WRITE_FD 1
+
+
+typedef BUFFER_UNIT_T buffer_unit_t;
+typedef struct {
+  buffer_unit_t *data;
+  size_t size;         /* size in bytes */
+  size_t bitpos;       /* bit offset from data  */
+} buffer_t;
+
+#define BUFFER_UNIT_SIZE (sizeof(buffer_unit_t))
+#define BUFFER_UNIT_BITS (BUFFER_UNIT_SIZE * 8)
+
+// Output buffer stack
+typedef struct {
+  buffer_t **data;
+  size_t capacity;
+  size_t size;
+} buf_stack;
+
+
+// State information
+struct state {
+  int num;
+  int accepting;
+};
+extern struct state state_table[];
+extern int state_count;
+
+// Program interface
+
+void printCompilationInfo();
+void init();
+int match(int phase, int start_state, char* buf, long length);
+void print_state_table();
